feat(front): change image preview styles
