+++ mylog/src/screens/map/AddLocationScreen.tsx
@@ -16,7 +16,8 @@ import MarkerSelector from '@/components/MarkerSelector';
 import AddLocationRightHeader from '@/components/AddLocationRightHeader';
 import CustomButton from '@/components/common/CustomButton';
 import DatePickerModal from '@/components/modal/DatePickerModal';
-import InputImagesViewer from '@/components/InputImagesViewer';
+import ImageUploader from '@/components/ImageUploader';
+import ImagesPreview from '@/components/ImagesPreview';
 import {useCreateMarker} from '@/hooks/queries/useMarker';
 import useGetAddress from '@/hooks/common/useGetAddress';
 import useDatePicker from '@/hooks/common/useDatePicker';
@@ -140,7 +141,10 @@ function AddLocationScreen({route, navigation}: AddLocationScreenProps) {
               onChangeDate={datePicker.handleChange}
               onConfirmDate={datePicker.handleConfirm}
             />
-            <InputImagesViewer />
+            <View style={styles.imagesViewer}>
+              <ImageUploader />
+              <ImagesPreview />
+            </View>
           </View>
         </ScrollView>
       </CustomKeyboardAvoidingView>
@@ -160,6 +164,9 @@ const styles = StyleSheet.create({
   inputContainer: {
     gap: 20,
     marginBottom: 20,
+  },
+  imagesViewer: {
+    flexDirection: 'row',
   },
 });
